Add Filtering for HR SELECTED STUDENTS and add refactoring for the filters functionality
